requirements.txt
# Core Dependencies
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Deep Learning & Audio
transformers>=4.30.0
librosa>=0.10.0
soundfile>=0.12.0
resampy>=0.4.2

# Model Components
speechbrain>=0.5.15
fairseq>=0.12.2
pesq>=0.0.4
pystoi>=0.3.3

# Training & Optimization
tensorboard>=2.13.0
wandb>=0.15.0
omegaconf>=2.3.0
hydra-core>=1.3.0

# Inference & Deployment
fastapi>=0.100.0
uvicorn>=0.23.0
onnx>=1.14.0
onnxruntime>=1.15.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
